pkg/auth: major test refactor
