+++ pkg/auth/authservice.go
@@ -1,7 +1,6 @@
 package auth
 
 import (
-	"crypto/ecdsa"
 	"errors"
 	"fmt"
 	"log"
@@ -37,43 +36,13 @@ var (
 	}
 )
 
-type TokenFactory struct {
-	Issuer        string
-	Audience      string
-	TokenValidity time.Duration
-	SigningKey    *ecdsa.PrivateKey
-}
-
-func (tf *TokenFactory) Create(
-	now time.Time,
-	subject string,
-) (*types.Token, error) {
-	expires := now.Add(tf.TokenValidity)
-	token := jwt.NewWithClaims(
-		jwt.SigningMethodES512,
-		jwt.StandardClaims{
-			Subject:   subject,
-			Audience:  tf.Audience,
-			Issuer:    tf.Issuer,
-			IssuedAt:  now.Unix(),
-			ExpiresAt: expires.Unix(),
-			NotBefore: now.Unix(),
-		},
-	)
-	t, err := token.SignedString(tf.SigningKey)
-	if err != nil {
-		return nil, fmt.Errorf("signing token: %w", err)
-	}
-	return &types.Token{Token: t, Expires: expires}, nil
-}
-
 type AuthService struct {
 	Creds         CredStore
 	Tokens        types.TokenStore
 	Notifications types.NotificationService
-	ResetTokens   ResetTokenFactory
+	ResetTokens   types.ResetTokenFactory
 	TokenDetails  TokenDetailsFactory
-	Codes         TokenFactory
+	Codes         types.TokenFactory
 	TimeFunc      func() time.Time
 }
 
@@ -241,7 +210,7 @@ func (as *AuthService) UpdatePassword(
 }
 
 func (as *AuthService) Exchange(code string) (*TokenDetails, error) {
-	var claims Claims
+	var claims types.Claims
 	if _, err := jwt.ParseWithClaims(
 		code,
 		&claims,
